requirements.txt
langchain
langchain_community
langchain-google-genai
python-dotenv
langchain_experimental
langchainhub
flask
faiss-cpu

-e .